requirements.txt
aiodns==3.2.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
asyncache==0.3.1
attrs==25.3.0
Brotli==1.1.0
cachetools==5.5.2
cffi==1.17.1
discord.py==2.5.2
dnspython==2.7.0
frozenlist==1.5.0
idna==3.10
motor==3.7.0
multidict==6.2.0
orjson==3.10.16
propcache==0.3.0
pycares==4.5.0
pycparser==2.22
pymongo==4.11.3
ruff==0.11.2
yarl==1.18.3
zstandard==0.23.0
